refactor: update search feature with latest chat implementation
