+++ packages/feeds-client/@react-bindings/contexts/StreamSearchContext.tsx
@@ -1,5 +1,5 @@
 import { createContext, useContext } from 'react';
-import type { SearchController } from '../../src/common/SearchController';
+import type { SearchController } from '../../src/common/search';
 
 export const StreamSearchContext = createContext<SearchController | undefined>(undefined);
 
@@ -11,7 +11,7 @@ export type StreamSearchContextProps = {
 };
 
 /**
- * Hook to access the nearest SearchController instance.
+ * Hook to access the nearest SearchController2 instance.
  */
 export const useSearchContext = () => {
   return useContext(StreamSearchContext);
