+++ packages/feeds-client/@react-bindings/wrappers/StreamSearch.tsx
@@ -1,9 +1,9 @@
 import { PropsWithChildren } from 'react';
 import { StreamSearchContext } from '../contexts/StreamSearchContext';
-import type { SearchController } from '../../src/common/SearchController';
+import type { SearchController } from '../../src/common/search';
 
 /**
- * The props for the StreamSearch component. It accepts a `SearchController` instance.
+ * The props for the StreamSearch component. It accepts a `SearchController2` instance.
  */
 export type StreamSearchProps = {
   searchController: SearchController | undefined;
