+++ packages/feeds-client/@react-bindings/wrappers/StreamSearchResults.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react';
 import { StreamSearchResultsContext } from '../contexts/StreamSearchResultsContext';
-import type { SearchSource } from '../../src/common/BaseSearchSource';
+import type { SearchSource } from '../../src/common/search';
 
 /**
  * The props for the StreamSearchResults component. It accepts a `SearchSource` instance.
