+++ packages/feeds-client/index.ts
@@ -4,10 +4,6 @@ export * from './src/gen/models';
 export * from './src/types';
 export * from './src/common/types';
 export * from './src/common/StateStore';
-export * from './src/common/BaseSearchSource';
-export * from './src/common/SearchController';
-export * from './src/common/ActivitySearchSource';
-export * from './src/common/UserSearchSource';
-export * from './src/common/FeedSearchSource';
+export * from './src/common/search';
 export * from './src/common/Poll';
 export * from './src/utils';
