+++ packages/feeds-client/src/common/search/index.ts
@@ -1,0 +1,6 @@
+export * from './SearchController';
+export * from './BaseSearchSource';
+export * from './ActivitySearchSource';
+export * from './FeedSearchSource';
+export * from './UserSearchSource';
+export * from './types';
