+++ sample-apps/react-native/ExpoTikTokApp/package.json
@@ -4,6 +4,7 @@
   "version": "1.0.0",
   "scripts": {
     "start": "expo start -c --dev-client",
+    "clean": "expo prebuild --clean",
     "android": "expo run:android",
     "ios": "expo run:ios",
     "web": "expo start --web",
